Converted dialog component to TypeScript
