--- src/components/dialog/index.js
@@ -1,1 +1,0 @@
-export { default } from './dialog.jsx';
